refactored coded, added inputField Class to make it reusable
