+++ src/contact.js
@@ -1,6 +1,6 @@
-/** @format */
+import { InputField } from './inputField';
 
-function contactSection() {
+export function contactSection() {
     // Create Elements
     let contactContainer = document.createElement('div');
     let contactSection = document.createElement('div');
@@ -8,19 +8,13 @@ function contactSection() {
     let contactInfo = document.createElement('p');
     let contactPhone = document.createElement('a');
     let contactForm = document.createElement('form');
-    let name = document.createElement('input');
-    let email = document.createElement('input');
+    let name = new InputField('name', 'text', 'First Name');
+    let email = new InputField('email', 'email', 'Email');
     let textArea = document.createElement('textarea');
     let button = document.createElement('button');
     // Give Elements ID
-    name.id = 'name';
-    email.id = 'email';
     textArea.id = 'feedback';
     // Set Attributes
-    name.setAttribute('type', 'text');
-    name.setAttribute('placeholder', 'First Name');
-    email.setAttribute('type', 'email');
-    email.setAttribute('placeholder', 'Email');
     textArea.setAttribute('cols', '30');
     textArea.setAttribute('rows', '10');
     textArea.setAttribute('placeholder', 'Give us your feedback!');
@@ -34,8 +28,8 @@ function contactSection() {
     contactContainer.classList.add('contact-container');
     contactSection.classList.add('contact');
     // Append Elements
-    contactForm.appendChild(name);
-    contactForm.appendChild(email);
+    contactForm.appendChild(name.createElement);
+    contactForm.appendChild(email.createElement);
     contactForm.appendChild(textArea);
     contactSection.appendChild(contactHead);
     contactSection.appendChild(contactInfo);
@@ -44,5 +38,3 @@ function contactSection() {
     contactContainer.appendChild(contactSection);
     return contactContainer;
 }
-
-export { contactSection };
